Add default for caseInsensitive...
This fix is necessary for campatibility for implementations using the
misspelled getCaseInsensitiveId method.  Note that if you reversed the
default, new implementation will fail since they don't implement the
misspelled version.  The solution is to have default implementations for
both versions.  This could, however, cause a stack overflow if no
implementation is provided.  To avoid this, we use introspection to see
if the default is used.

Note that this changes the behavior of a missing interface from a
compile time error to a run-time error.